moved MockOutfists.js and updated
